Generate ranks model

+++ db/migrate/20220420152205_add_rank_to_products.rb
@@ -1,5 +1,5 @@
 class AddRankToProducts < ActiveRecord::Migration[7.0]
   def change
-    add_column :products, :rank, :decimal, default: 0.00
+    add_column :products, :rank, :integer, default: 0
   end
 end
